Create business-insider-bookmarklet.js
This will bypass the paywall for any (Business)Insider.com articles and load the content in a new browser popup window. It is formatted for use as a bookmarklet: just save this code in the URL field of a new bookmark/favorite in your browser (the name field can be whatever you like - e.g., Insider.com article paywall bypass). When viewing any Insider.com article page, just select the bookmark and it will instantly launch the new popup with the article content.